+++ src/input2/keyboard.js
@@ -1,7 +1,0 @@
-
-export class KeyboardAgent {
-
-  constructor(params) {
-    
-  }
-}